+++ options/options_parser.cc
@@ -574,7 +574,7 @@ bool AreEqualOptions(
       CompactionOptionsFIFO rhs =
           *reinterpret_cast<const CompactionOptionsFIFO*>(offset2);
       if (lhs.max_table_files_size == rhs.max_table_files_size &&
-          lhs.ttl == rhs.ttl && lhs.allow_compaction == rhs.allow_compaction) {
+          lhs.allow_compaction == rhs.allow_compaction) {
         return true;
       }
       return false;